timeremaining is now float64

+++ go/models/bid.go
@@ -14,7 +14,7 @@ type BidValue struct {
 type BidRequest struct {
 	PlayerId      string        `json:"playerId"`
 	PlayerIds     []string      `json:"playerIds"`
-	TimeRemaining int           `json:"timeRemaining"`
+	TimeRemaining float64       `json:"timeRemaining"`
 	Cards         []string      `json:"cards"`
 	BidHistory    []interface{} `json:"bidHistory"`
 	BidState      BidState      `json:"bidState"`
