+++ go/models/play.go
@@ -9,7 +9,7 @@ type Team struct {
 type PlayRequest struct {
 	PlayerId      string          `json:"playerId"`
 	PlayerIds     []string        `json:"playerIds"`
-	TimeRemaining int             `json:"timeRemaining"`
+	TimeRemaining float64         `json:"timeRemaining"`
 	Teams         []Team          `json:"teams"`
 	Cards         []string        `json:"cards"`
 	BidHistory    [][]interface{} `json:"bidHistory"`
